Two cutscene fixes
- Fixed Trogdor sprite leaving the game window in cutscene 1 by rendering an additional black rect where Trogdor goes
- Adjusted archer starting position in cutscene 11; this was accurate to the original game, but since this version doesn't have the original's rounded border, the archer was sticking out too far from the edge of the screen

+++ Trogdor-Reburninated/src/classes.cpp
@@ -660,7 +660,7 @@ void GameManager::setArcherFrequency() {
 			}
 			break;
 		case 1: // scale from (level 1 = 20) to (level 31 = 400)
-			archerFrequency = ceil((12.6666 * level) + (7.3333));
+			archerFrequency = (Uint16)ceil((12.6666 * level) + (7.3333));
 			break;
 		case 2:
 			archerFrequency = 40;
